requirements.txt
## Base URL of the Python Package Index.
## This should point to a repository compliant with PEP 503 (the simple
## repository API) or a local directory laid out in the same format.
-i https://pypi.python.org/simple/
# -i https://pypi.tuna.tsinghua.edu.cn/simple/


## If a URL or path to a html file, then parse for links to archive
## such as sdist (.tar.gz) or wheel (.whl) files. If a local path or
## file:// URL that's a directory, then look for archives in the directory
## listing. Links to VCS project URLs are not supported.
-f https://download.pytorch.org/whl/torch_stable.html
# -f https://s3.amazonaws.com/pytorch/whl/torch_stable.html
# -f https://mirror.sjtu.edu.cn/pytorch-wheels/torch_stable.html


## All dependencies for installation
alaas==0.2.0
azureml-datadrift==1.45.0
matplotlib==3.5.3
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
scipy==1.8.0
scikit-learn==1.1.2
tqdm==4.64.0
torchmetrics==0.9.3

## Cuda version >= 11.3 [Recommend]
torch==1.11.0+cu113
torchvision==0.12.0+cu113


## Cuda version < 11.3
# torch==1.11.0+cu102
# torchvision==0.12.0+cu102
